Fix excessive memory usage
I experienced this reproducibly when spacemacs reached the stage where it shows
the message in the mode line starting like this:

"Downloading stable ELPA repository: "

Without this change I was unable to start spacemacs and in fact had to reboot my
machine multiple times. Emacs 25.1.1, Ubuntu 17.04, no swap configured.

Obviously this needs cleaning up a bit...

Also fixed an extraneous 'kkj' which I assume is a typo (actually I thought this
was my typo but turns out to be syl20bnr himself :-)